+++ e2e/commands/commit.e2e.js
@@ -94,6 +94,18 @@ describe('bit tag command', function () {
         const tagWithExisting = () => helper.commitComponent('components/exact', 'message', '-f --exact_version 5.5.5');
         expect(tagWithExisting).to.throw('the version 5.5.5 already exists for components/exact');
       });
+      it('Should print same output for flaged tag and non flaged tag', () => {
+        helper.reInitLocalScope();
+        helper.createFile('components', 'major.js');
+        helper.addComponent('components/major.js');
+        const majorOutput = helper.commitComponent('components/major', 'message', '--major');
+        helper.reInitLocalScope();
+        helper.createFile('components', 'major.js');
+        helper.addComponent('components/major.js');
+        const nonFlagedCommit = helper.commitComponent('components/major');
+        expect(majorOutput).to.contain('1 components tagged | 1 added, 0 changed, 0 auto-tagged');
+        expect(nonFlagedCommit).to.contain('1 components tagged | 1 added, 0 changed, 0 auto-tagged');
+      });
     });
     describe('tag all components', () => {
       before(() => {